api :: todo

+++ api/controllers/todoPagesController.js
@@ -1,5 +1,6 @@
-const { TODO_PAGE_NOT_FOUND } = require("../constants");
+const { TODO_PAGE_NOT_FOUND, USER_NOT_FOUND } = require("../constants");
 const TodoPage = require("../models/TodoPage");
+const Todo = require("../models/Todo");
 const User = require("../models/User");
 const { sendResponseIfExists } = require("../utils/sendResponseIfExists");
 
@@ -33,10 +34,12 @@ module.exports.createTodoPage = (req, res) => {
     { $push: { todoPages: todoPage._id } }
   ).then((result) => {
     console.log("RES:", result);
-    todoPage
-      .save()
-      .then((savedTodoPage) => res.send(savedTodoPage))
-      .catch((err) => res.status(400).send(err));
+    if (result) {
+      todoPage
+        .save()
+        .then((savedTodoPage) => res.send(savedTodoPage))
+        .catch((err) => res.status(400).send(err));
+    } else res.status(400).send(USER_NOT_FOUND);
   });
 };
 
@@ -44,9 +47,9 @@ module.exports.updateTodoPage = (req, res) => {
   const { tpId } = req.params;
   TodoPage.findByIdAndUpdate(tpId, req.body, { new: true, runValidators: true })
     .then((todoPage) =>
-      sendResponseIfExists(todoPage, res, TODO_PAGE_NOT_FOUND, 404)
+      sendResponseIfExists(todoPage, res, TODO_PAGE_NOT_FOUND, 400)
     )
-    .catch((err) => res.status(400).send(err));
+    .catch((err) => res.status(404).send(err));
 };
 
 module.exports.deleteTodoPage = (req, res) => {
@@ -56,20 +59,16 @@ module.exports.deleteTodoPage = (req, res) => {
     { $pull: { todoPages: tpId } }
   ).then((result) => {
     console.log("RES:", result);
-    TodoPage.findByIdAndDelete(tpId)
-      .then((deleted) =>
-        sendResponseIfExists(deleted, res, TODO_PAGE_NOT_FOUND, 400)
-      )
-      .catch((err) => res.json(err));
+    if (result) {
+      TodoPage.findByIdAndDelete(tpId)
+        .then((deleted) => {
+          if (deleted && deleted.todos) {
+            Todo.deleteMany({ id: { $in: deleted.todos } })
+              .then((deletedTodosInfo) => res.send(deleted))
+              .catch((err) => res.status(404).send(err));
+          } else res.status(400).send(TODO_PAGE_NOT_FOUND);
+        })
+        .catch((err) => res.json(err));
+    } else res.status(400).send(USER_NOT_FOUND);
   });
-};
-
-module.exports.deleteAll = (req, res) => {
-  User.findOneAndUpdate({ _id: req.body.id }, { $set: { todoPages: [] } })
-    .then((result) =>
-      TodoPage.deleteMany({ id: { $in: result.todoPages } })
-        .then((result2) => res.send(result2))
-        .catch((err) => res.status(404).send(err))
-    )
-    .catch((err) => res.status(404).send(err));
 };
